# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Saving Grains API
  description: |
    The Saving Grains API provides endpoints for user authentication, user management, and wallet request functionality. 

    ## Authentication
    Most endpoints require authentication via JWT token passed in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ## Error Handling
    All API errors follow a consistent format with an `error` field containing the error message.
  version: 1.0.0
  contact:
    name: Saving Grains Support
    email: support@savinggrains.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.savinggrains.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message description
          example: "Invalid credentials"
    LoginRequest:
      type: object
      required:
        - phone_number
        - password
      properties:
        phone_number:
          type: string
          description: User's registered phone number
          example: "+233123456789"
        password:
          type: string
          description: User's password
          example: "securePassword123"
    VerifyOTPRequest:
      type: object
      required:
        - phone_number
        - otp
      properties:
        phone_number:
          type: string
          description: User's registered phone number
          example: "+233123456789"
        otp:
          type: string
          description: 6-digit OTP received via SMS
          example: "123456"
    AuthValidateReq:
      type: object
      required:
        - phoneNumber
      properties:
        phoneNumber:
          type: string
          description: International phone number to validate
          example: "+233123456789"
        countryCode:
          type: string
          description: Optional ISO country code (e.g., GH)
          example: "GH"
    AuthValidateRes:
      type: object
      properties:
        frequency:
          type: string
          description: Frequency classification for the user
          example: "FIRST_TIMER"
          enum: ["FIRST_TIMER", "FREQUENT", "STALE"]
        firstName:
          type: string
          description: User's first name (if known)
          example: "John"
...
paths:
  /auth/login:
    post:
      summary: Authenticate user and send OTP
      description: Validates user credentials and sends a 6-digit OTP via SMS for two-factor authentication
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OTP sent via SMS"
        "400":
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Phone number and password are required"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Invalid credentials"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "User not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Internal server error"
  /auth/verify-otp:
    post:
      summary: Verify OTP and complete login
      description: Verifies the OTP sent during login and returns a JWT token for authenticated requests
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOTPRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  token:
                    type: string
                    description: JWT token for authenticated requests (expires in 7 days)
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "400":
          description: Bad request - missing fields or invalid/expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                missing_fields:
                  value:
                    error: "Phone number and OTP required"
                invalid_otp:
                  value:
                    error: "Invalid OTP"
                expired_otp:
                  value:
                    error: "OTP expired"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "User not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Server error"
  /auth/validate:
    post:
      summary: Validate a phone number and decide next auth step
      description: Checks whether a phone number belongs to an existing user and returns a frequency indicator and firstName for client flow decisions.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthValidateReq"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthValidateRes"
        "400":
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "phoneNumber is required"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Internal server error"
...
security:
  - bearerAuth: []